Discovery handler
Summary:
This adds a handler for a `/discover` route. Its only purpose is to return 200 so that other drones in the network can properly recognize the Skyserve instance on this drone.

Another library will discover devices on the network and hit this route to check healthiness of the server.

Reviewers: andrew

Reviewed By: andrew

Maniphest Tasks: T50

Differential Revision: https://code.skynet.engineering/D20