Active missions handler
Summary:
This changes adds a path `/active-missions` to allow a client to query all active missions running on the target drone. Missions are self-registered into Redis on initialization.

Eventual use case is to allow the GCS to query all drones in the network and the missions each is running. Some frontend can then select drone and target mission to send commands to.

Client return format is of the shape

```
{"missions": [{"mission_id": ..., "port": ...}, ...]}
```

Test Plan: Manual verification with starting and stopping missions with v2 of the Skymission interface

Reviewers: #skynet

Maniphest Tasks: T128

Differential Revision: https://code.skynet.engineering/D85